refactor: Switch to a behaviour to update the colors when queries change
